requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.3
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.5.1
mypy_extensions==1.1.0
narwhals==1.44.0
networkx==3.5
numpy==1.26.4
openai==1.92.2
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
playwright==1.53.0
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
storage3==0.12.0
streamlit==1.46.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.1.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
